backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Celery & Task Queue
celery[redis]==5.3.6
redis==5.0.1

# PDF Processing
pikepdf==8.11.2

# File handling & Security
python-magic==0.4.27
aiofiles==23.2.1
clamd==1.0.2

# Database (SQLite for job metadata)
sqlalchemy==2.0.25
alembic==1.13.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# HTTP & Webhooks
httpx==0.26.0

# Utilities
tenacity==8.2.3
shortuuid==1.0.11

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.0.0

# HTTP client (for healthcheck)
requests==2.31.0










